tidy up: put reusable pydantic models in models.py and export them
