+++ pgscatalog.core/src/pgscatalog/core/lib/__init__.py
@@ -1,10 +1,8 @@
 from ._config import Config
-from .allele import Allele
 from .effecttype import EffectType
 from .genomebuild import GenomeBuild
 from .catalogapi import ScoreQueryResult, CatalogQuery, CatalogCategory
 from .scorefiles import ScoringFiles, ScoringFile, NormalisedScoringFile
-from .catalogscorevariant import CatalogScoreVariant
 from .scorevariant import ScoreVariant
 from .targetvariants import TargetVariants, TargetVariant, TargetType
 from ._relabel import RelabelArgs, relabel, relabel_write
@@ -29,6 +27,8 @@ from .pgsexceptions import (
     GenomesNotFound,
     SamplesheetFormatError,
 )
+
+from . import models
 
 __all__ = [
     "BasePGSException",
@@ -66,7 +66,6 @@ __all__ = [
     "relabel_write",
     "effect_type_keyfunc",
     "chrom_keyfunc",
-    "CatalogScoreVariant",
     "EffectType",
-    "Allele",
+    "models",
 ]
